Migrate to `nixos-sf-ssh-auth` ssh auth dir
Adapted on config factory init code to make sure factory user
does not lock himself out of the device by making user of the
new `nixos-sf-ssh-auth` ssh auth dir structure and a special
`factory-installed-by` per device field.

It also no longer authorize the factory user to all devices but
instead only add the factory user as a ssh user to the auth dir
and merely set the `factory-installed-by` array field (added to the
device state) so that its sole member is the current factory user.

The previous make the assumption a particular ssh auth setup
was used to grant special ssh priviledges to users in the
device's `factory-installed-by` listing.